ref: reorganize component structure according to methodology
