+++ client/src/components/organisms/bet/UserBetCard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Card, CardContent, CardFooter } from "../ui/card";
+import { Card, CardContent, CardFooter } from "@/components/ui/card";
 import BetCardHeader from "@/components/molecules/Bet/BetCardHeader";
 import "@/css/accordion-overrides.css";
 import BetInfoSummary from "@/components/molecules/Bet/BetInfoSummary";
