+++ client/src/pages/BetsPage.jsx
@@ -1,5 +1,5 @@
-import BetsTabs from "@/components/organisms/BetsTabs";
-import Navbar from "@/components/organisms/Navbar";
+import BetsTabs from "@/components/organisms/bet/BetsTabs";
+import Navbar from "@/components/organisms/navigation/Navbar";
 import React, { useEffect, useState } from "react";
 
 const BetsPage = () => {
